+++ apis/v1/conference/page.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+import { TopConferenceList } from '@/types/research';
+
+export const getTopConferenceList = () => getRequest<TopConferenceList>('/v1/conference/page');
