+++ app/[locale]/research/labs/page.tsx
@@ -1,4 +1,4 @@
-import { getResearchLabs } from '@/apis/research';
+import { getResearchLabs } from '@/apis/v2/research/lab';
 import { OrangeButton } from '@/components/common/Buttons';
 import LoginVisible from '@/components/common/LoginVisible';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
