+++ app/[locale]/research/top-conference-list/page.tsx
@@ -1,6 +1,6 @@
 import { getTranslations } from 'next-intl/server';
 
-import { getTopConferenceList } from '@/apis/research';
+import { getTopConferenceList } from '@/apis/v1/conference/page';
 import ConferenceListTable from '@/app/[locale]/research/top-conference-list/ConferenceListTable';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { getMetadata } from '@/utils/metadata';
